Extract controller vars into object
According to Sandi Metz's rules for developers, we should only
instantiate one instance variable to pass to a view in each controller
action.  The dashboard view has so many different data points to display
that we need to extract this complexity into its own Plain Old Ruby
Object that we can test better.