Remove payments table
Adjust Dashboard to not care about how much has been paid

Getting rid of the concept of payments means this app will
just be about the amount of expenses incurred - it will not
worry about the amount reimbursed. This is a good way to
limit the scope of the app and keep things as simple as
possible.