Change namespace from User to CurrentUser
`User` conflicted with an ActiveRecord model class.

It is better to use CurrentUser, which is solely a module name and
does not conflict with any classes.

A lot of files had to be altered to make this change. Is this a sign
of something wrong with the design, like a lot of coupling?  Why did
so many places need to be changed?